chore: cleanup router
